Separated OutputManager methods into two OutputWriter classes
