+++ include/PaleoSimH5OutputWriter.hh
@@ -1,0 +1,11 @@
+#ifndef PALEOSIMH5OUTPUTWRITER_HH
+#define PALEOSIMH5OUTPUTWRITER_HH
+
+class PaleoSimOutputManager;
+
+class PaleoSimH5OutputWriter {
+public:
+    static void Write(PaleoSimOutputManager& manager);
+};
+
+#endif
