+++ include/PaleoSimRootOutputWriter.hh
@@ -1,0 +1,11 @@
+#ifndef PALEOSIMROOTOUTPUTWRITER_HH
+#define PALEOSIMROOTOUTPUTWRITER_HH
+
+class PaleoSimOutputManager;
+
+class PaleoSimRootOutputWriter {
+public:
+    static void Write(PaleoSimOutputManager& manager);
+};
+
+#endif
